Correct test for #1775
